+++ src/test/kotlin/me/bossm0n5t3r/leetcode/excelSheetColumnNumber/ExcelSheetColumnNumberTest.kt
@@ -1,10 +1,10 @@
-package excelSheetColumnNumber
+package me.bossm0n5t3r.leetcode.excelSheetColumnNumber
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class ExcelSheetColumnNumberTest {
-    private val excelSheetColumnNumber = ExcelSheetColumnNumber()
+    private val excelSheetColumnNumber = ExcelSheetColumnNumber.Solution()
 
     data class ExcelSheetColumnNumberTestData(
         val columnTitle: String,
